+++ utils/strings.js
@@ -1,0 +1,10 @@
+const pembangkitNames = {
+  pltu: 'Pembangkit Listrik Tenaga Uap (PLTU)',
+  pltg: 'Pembangkit Listrik Tenaga Gas (PLTG)',
+  pltmg: 'Pembangkit Listrik Tenaga Mesin Gas (PLTMG)',
+  pltd: 'Pembangkit Listrik Tenaga Diesel (PLTD)',
+};
+
+module.exports = {
+  pembangkitNames,
+};
